Move KPI metrics factory and KPI metrics config code to metrics/metrics
