+++ metrics/metrics/src/main/java/io/helidon/metrics/KeyPerformanceIndicatorMetricsConfig.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.helidon.webserver;
+package io.helidon.metrics;
 
 /**
  * Configuration for KPI metrics.
