+++ metrics/metrics/src/main/java/io/helidon/metrics/KeyPerformanceIndicatorMetricsConfigImpl.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.helidon.webserver;
+package io.helidon.metrics;
 
 class KeyPerformanceIndicatorMetricsConfigImpl implements KeyPerformanceIndicatorMetricsConfig {
